+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network.WebrtcAdapter/index.ts
@@ -1,0 +1,1 @@
+export * from './NetworkAdapter';
