+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network.WebrtcAdapter/t.ts
@@ -1,0 +1,1 @@
+export type { WebrtcNetworkAdapter } from './NetworkAdapter';
