+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network/common.ts
@@ -1,1 +1,2 @@
 export * from '../common';
+export { WebrtcNetworkAdapter } from '../Store.Network.WebrtcAdapter';
