+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network/index.ts
@@ -1,2 +1,2 @@
-export { WebrtcNetworkAdapter } from './NetworkAdapter';
+export { WebrtcNetworkAdapter } from '../Store.Network.WebrtcAdapter';
 export { WebrtcStore } from './Store';
