+++ code/ext/ext.lib.automerge.webrtc/src/Store.Network/t.Store.ts
@@ -1,5 +1,4 @@
 import type { t } from './common';
-export type { WebrtcNetworkAdapter } from './NetworkAdapter';
 
 type Uri = string;
