+++ code/ext/ext.lib.automerge.webrtc/src/types.ts
@@ -1,4 +1,5 @@
 export type * from './Store.Network/t';
+export type * from './Store.Network.WebrtcAdapter/t';
 
 export type * from './ui/ui.Info/t';
 export type * from './ui/ui.NetworkConnection/t';
